faiss-bridge/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
faiss-cpu==1.8.0
numpy==1.26.4
python-dotenv==1.0.1
openai==1.35.7
orjson==3.10.3
python-multipart==0.0.9

